Allow gdb to read the target registers and start the target.
In order to read this milestone, which allows gdb to actually finish
its handshaking procedure with fetproxy, this patch completes the link
between GdbClient and FetModule.

In order to keep the interface to GdbClient somewhat generic, allowing
for its use with other targets etc, it gets fed with a struct full of
callbacks that it can use to control the target.  At the moment there
are two callbacks that control the target:

 * read_registers - Ask the target to read all registers
 * cont - Ask the target to continue execution

When FetModule has received the UIF's response to the given command,
it calls gdb_client_command_complete.  This allows GdbClient to
process the new target state and relay any data onto gdb where
appropriate.

If GdbClient has any further queued requests to send to the FetModule,
it will send them on after receiving its response.  At the moment, I
am unsure whether there can ever be more than one request queued from
gdb -- and may remove this incoming queue in the future.

GdbClient remembers what the last request it sent to FetModule was,
allowing it to perform the correct operation when
gdb_client_command_complete is called.